Finished the script
All conversions steps I found have been added and tested. The script
isn't "optimized" and took a good 20 minutes to run on 300-400 MB with
one core on a Core i3.

The script might have a quirk or weakness here and there, but it is fine
for the majority of the text afaic.

A file with sample text is included as well as an image that illustrates
differences in output between fonts.